requirements.txt
requests==2.32.3
pandas==2.2.2
pyarrow==17.0.0
mwparserfromhell==0.6.6
python-dateutil==2.9.0.post0
pytest==8.3.2
